requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
jinja2>=3.1
sqlmodel==0.0.22
SQLAlchemy==2.*
python-multipart>=0.0.9
passlib[argon2]>=1.7
argon2-cffi>=23.1
itsdangerous>=2.1
python-dotenv>=1.0
httpx